requirements.txt
email-validator==2.1.0.post1
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
requests==2.31.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1